client/requirements.txt
imutils
numpy
opencv-python
python-decouple
python-telegram-bot
matplotlib
tensorflow-gpu